[terminal_renderer] Improving rendering of curly underline and some refactors along the lines.
